+++ src/frontend/reducer/notesReducer.js
